tourbook: Fix display of stale data
When switching to botches, ascends or projects stale data could be
displayed, if one of these lists was empty.
This is fixed by always displaying an empty list, if no ascend are
available instead of skipping the redraw event.

Additionally, the heading for ascends without a date is now empty
instead of "0".

Signed-off-by: Fabian Kantereit <fabian.kantereit@tsaro.io>