Add button for database reset to preferences (#242)
Add a reset database button ("Auf Werkseinstellungen zurücksetzen") to the preferences view. After a confirmation dialog it clears the database and tour book and also resets the custom settings to the default values. After deleting a success toast is shown.

For future bugs it might be helpful to give the user the opportunity to reset all databases in the app.

Resolves #242